+++ testData/org/elixir_lang/documentation/defdelegate_quick_doc/delegate_declaration.ex
@@ -1,0 +1,12 @@
+defmodule DelegateTarget do
+  @moduledoc "DelegateTarget module."
+  @doc "Merges two maps."
+  def merge(map1, map2) do
+    Map.merge(map1, map2)
+  end
+end
+
+defmodule Delegator do
+  @moduledoc "Delegator module."
+  defdelegate mer<caret>ge(map1, map2), to: DelegateTarget
+end